+++ internal/app/storage/storager.go
@@ -2,7 +2,7 @@ package storage
 
 type IStorager interface {
 	GetItem(itemID string) (string, error)
-	AddItem(itemID string, value string, userID uint64)
+	AddItem(itemID string, value string, userID uint32)
 	GetLastElementID() string
-	GetItemsByUserID(userID uint64) []UserURLs
+	GetItemsByUserID(userID uint32) []UserURLs
 }
